colexechash: switch to operating with uint32 values
This commit switches the hash table to operate on uint32 hash values
rather than uint64. This switch allows us to reduce the memory
footprint as well as to actually speed up the computation. Full
microbenchmarks are [here](https://gist.github.com/yuzefovich/8a672586343a903f2669fa883c67c265),
but the gist is as follows:
- hash aggregator has speedup of 10-20% and reduced mem usage of 5-40%
- unordered distinct has speedup of 2-6% and reduced mem usage of 2-30%
- hash joiner has speedup of 2-25% and reduced mem usage of 4-10%

Difference on TPCH queries (average over 100 runs):
```
Q1:	before: 3.32s	after: 3.16s	 -4.93%
Q2:	before: 3.42s	after: 3.15s	 -8.08%
Q3:	before: 2.68s	after: 2.39s	 -10.88%
Q4:	before: 1.72s	after: 1.64s	 -4.46%
Q5:	before: 2.45s	after: 2.35s	 -4.22%
Q6:	before: 4.56s	after: 4.49s	 -1.49%
Q7:	before: 5.69s	after: 5.57s	 -2.16%
Q8:	before: 1.10s	after: 1.09s	 -0.66%
Q9:	before: 5.69s	after: 5.80s	 2.00%
Q10:	before: 2.09s	after: 1.98s	 -5.17%
Q11:	before: 0.95s	after: 0.91s	 -4.13%
Q12:	before: 4.26s	after: 4.28s	 0.57%
Q13:	before: 1.34s	after: 1.26s	 -6.20%
Q14:	before: 0.44s	after: 0.45s	 2.28%
Q15:	before: 2.47s	after: 2.47s	 -0.13%
Q16:	before: 0.93s	after: 0.88s	 -4.68%
Q17:	before: 0.25s	after: 0.24s	 -2.76%
Q18:	before: 2.18s	after: 2.12s	 -2.48%
Q19:	before: 0.51s	after: 0.49s	 -2.79%
Q20:	before: 9.80s	after: 9.53s	 -2.70%
Q21:	before: 4.96s	after: 4.85s	 -2.26%
Q22:	before: 0.58s	after: 0.57s	 -1.67%
```

The rationale for why this change is not breaking is that we don't expect
to need more than 4 billion buckets. For the hash table that would mean
that about 4 billion rows are buffered in memory which we should never
reach given the workmem limit that kicks in at around million of tuples
in the hash table.

Note that we still internally compute the hash of stuff using uint64,
but then later when we assign it to return outside of the hash functions
we trim it down to uint32.

However, in order to not change the routing behavior of the hash router
(which uses the tuple hash distributor that is also powered by the hash
utilities) we need to preserve uint64 usage in that case. We do so by
duplicating a couple of helper functions and making another function
generic. Note that we couldn't make `initHash` because we use
concretely-typed slice for fast zeroing out, but we could have made
`rehash` generic, however, that currently incurs a small performance
penalty, so we choose to duplicate the generated code there.

Epic: None

Release note: None

+++ pkg/sql/colexec/colexecagg/aggregators_util.go
@@ -40,7 +40,7 @@ type NewAggregatorArgs struct {
 		// HashTableNumBuckets if positive will override the initial number of
 		// buckets to be used by the hash table (if this struct is passed to the
 		// hash aggregator).
-		HashTableNumBuckets uint64
+		HashTableNumBuckets uint32
 	}
 }
